fix(etcd): use different api to get single key or range of keys from etcd (#336)
The main difference between the `get` and `readdir` functions in the
etcd client lies in whether or not to pass the `range_end` parameter.


https://github.com/api7/lua-resty-etcd/blob/ea0f4abe9cb3c00b6c6e6845f7a36ec2db27cf63/lib/resty/etcd/v3.lua#L1267-L1282


https://github.com/api7/lua-resty-etcd/blob/ea0f4abe9cb3c00b6c6e6845f7a36ec2db27cf63/lib/resty/etcd/v3.lua#L1315

When querying a single key, passing the `range_end` parameter will bring
about the following issues:
1. Performance problem: Passing the `range_end` means executing batch
queries, which is obviously less efficient than querying a single key.
2. In our control plane, we use the https://github.com/k3s-io/kine to
replace etcd. When querying a single key with kine, if we pass the
`range_end`, it will result in being unable to find the desired key.
This is an issue specific to kine itself. However, since etcd's official
client does not set the `range_end` parameter when querying a single
key, I prefer for apisix to behave more like the official client when
acting as an etcd client.

Signed-off-by: Nic <qianyong@api7.ai>