fix: failing plugin_metadata schema check at init (#429)
https://project.feishu.cn/zhiliu-ee/story/detail/4293680254

## Description
The schema check
[here](https://github.com/api7/api7-ee-3-gateway/blob/9f4ba1ef9e894978e7f39969a856212005281ed0/apisix/plugin.lua#L1078)
always fails because when adding plugin fields via plugin_metadata, not
all fields will be passed. So this PR removes this schema check. Its no
issue because schema check already happens when plugin is originally
added via route/global_plugin.

### Why wasn't this caught earlier?
The schema check function expects a name but name was nil
[here](https://github.com/api7/api7-ee-3-gateway/blob/9f4ba1ef9e894978e7f39969a856212005281ed0/apisix/plugin.lua#L920)
which caused this function to assume that this is a disabled plugin and
it returned
[here](https://github.com/api7/api7-ee-3-gateway/blob/9f4ba1ef9e894978e7f39969a856212005281ed0/apisix/plugin.lua#L931)
skipping the schema check anyways.