revert: add back removed schema check from plugin_metadata (#472)
In this PR https://github.com/api7/api7-ee-3-gateway/pull/429, the
plugin_metadata schema check which takes place at initialization time
was removed which was not necessary. Because just making sure that [id
is present
here](https://github.com/api7/api7-ee-3-gateway/pull/429/files#diff-33b785623e329abe054a672034b69660658b76e8035eef101169832c4e0b4ba1R217)
would have solved the issue. This PR adds back the removed schema check
and test that confirm it.

This validation was previously being skipped completely because id was
nil and even in our tests there are cases where plugin metadata is added
which doesn't conform to the schema.

So in our tests and everywhere we need to make sure the schema is
conformed. (This will be done in another PR)

Even in our tests we didnt provide proper schema conforming object
for plugin metadata
Because this was in our tests deliberately , I was mistaken to think of
it as not a bug

---------

Signed-off-by: Ashish Tiwari <ashishjaitiwari15112000@gmail.com>