chore: remove soap plugin (#480)
https://project.feishu.cn/zhiliu-ee/story/detail/4542494130

Compared with the competitor Kong Gateway, the current API7 Gateway
component has a larger image size because it contains Python and other
dependencies;

Currently, no production customers are using SOAP, so there is no need
to include this dependency;

For customers with this requirement, packages containing SOAP can be
delivered in a targeted manner, or a more decoupled solution can be
adopted. For example: run the logic for processing conversion in an
external service.

After removing the dependency, it is bound to affect the API and UI. My
idea is: Does the API have the ability to sense whether it can handle
the activation of the SOAP plug-in? (So I chose to require API design
and PRD);

In addition to the above requirements, it is also necessary to check
whether there are similar situations in GW? It needs to be checked and
discussed one by one.

The volume should be controlled to be similar to the size of the open
source apisix. Maybe an rpm may be needed to solve it?